requirements.txt
django==3.2.8
environs==9.3.4
django-phonenumber-field==5.1.0
phonenumbers==8.12.21
django-import-export==2.6.1
django-admin-shortcuts==2.0.0
pandas==1.3.4